add wandb logger
