+++ .gitignore
@@ -104,6 +104,7 @@ ENV/
 data/
 input/
 saved/
+wandb/
 
 # editor, os cache directory
 .vscode/
